Account for no audio provider in AudioPlaybackPanel so opening BRSTM doesn't crash in Mono
(cherry picked from commit e6ccb97ec97b68e72eeb1a7621783153e6ce9b9c)

+++ BrawlLib/SSBB/ResourceNodes/RSAR/File Types/RSTMNode.cs
@@ -103,7 +103,7 @@ namespace BrawlLib.SSBB.ResourceNodes
             _blockSize = part1->_blockSize;
             _bps = part1->_bitsPerSample;
 
-            int offset = ((int)(Header->DATAData->Data) - (int)(Header));
+            int offset = Header->DATAData->Data - Header;
             if (offset < WorkingUncompressed.Length)
             {
                 _audioSource = new DataSource(Header->DATAData->Data, WorkingUncompressed.Length - offset);
